feat(oog_call): align empty account check with callop
